# admin-backend/docs/swagger.yaml
definitions:
...
  admin_proto.RespAdminRoleInfoData:
    properties:
      createAdminId:
        description: 创建人ID
        type: integer
      createAdminName:
        description: 创建人名称
        type: string
      createdAt:
        description: 创建时间
        type: string
      describe:
        description: 角色描述
        type: string
      id:
        description: 角色ID
        type: integer
      isEnabled:
        description: 启用状态
        type: boolean
      modifyAdminId:
        description: 修改人ID
        type: integer
      modifyAdminName:
        description: 修改人名称
        type: string
      name:
        description: 角色名称
        type: string
      permissionIds:
        items:
          type: integer
        type: array
      permissions:
        items:
          $ref: '#/definitions/admin_proto.RolePermissionItem'
        type: array
      updatedAt:
        description: 更新时间
        type: string
    type: object
  admin_proto.RespAdminRoleListData:
    properties:
      list:
        items:
          $ref: '#/definitions/admin_proto.RoleItem'
        type: array
      total:
        type: integer
    type: object
...
  admin_proto.RoleItem:
    properties:
      createAdminId:
        description: 创建人ID
        type: integer
      createAdminName:
        description: 创建人名称
        type: string
      createdAt:
        type: string
      describe:
        description: 角色名称
        type: string
      id:
        description: 角色ID
        type: integer
      isEnabled:
        description: 是否启用
        type: boolean
      name:
        description: 角色名称
        type: string
      updatedAt:
        type: string
    type: object
  admin_proto.RolePermissionItem:
    properties:
      permissionId:
        description: 权限ID
        type: integer
      permissionKey:
        description: 权限键名
        type: string
      permissionName:
        description: 权限名称
        type: string
      permissionType:
        description: 权限类型
        type: string
      permissionTypeText:
        description: 权限类型
        type: string
      roleId:
        description: 权限ID
        type: integer
    type: object
  code.Message:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/code_proto.ErrorCode'
        description: 错误码
      data: {}
      msg:
        description: 错误码描述
        type: string
      reason:
        description: 错误具体信息
        type: string
    type: object
  code_proto.ErrorCode:
    enum:
    - 0
    - 1
    - 100001
    - 100002
    - 100003
    - 200001
    - 200002
    - 200003
    - 200004
    - 200005
    - 300001
    - 300002
    - 300003
    - 400001
    - 400002
    - 400003
    - 500001
    - 500002
    - 500003
    - 500004
    - 600001
    - 600002
    - 600003
    - 600004
    - 600005
    - 700001
    - 800001
    type: integer
    x-enum-comments:
      ErrorCode_AdminAccountInvalid: 账号已被封禁
      ErrorCode_AdminAccountNotExist: 账号不存在或已被删除
      ErrorCode_AdminAccountPasswordInvalid: 账号或密码错误
      ErrorCode_AdminApiKeyExist: 接口键名已存在
      ErrorCode_AdminApiNameExist: 接口名称已存在
      ErrorCode_AdminApiNotExist: 接口不存在货已失效
      ErrorCode_AdminApiPathExist: 接口路径已存在
      ErrorCode_AdminMenuNotExist: 菜单不存在
      ErrorCode_AdminPermissionExist: 该菜单下权限已存在
      ErrorCode_AdminPermissionKeyExist: 权限键名已存在
      ErrorCode_AdminPermissionKeyNeed: 权限键名不能为空
      ErrorCode_AdminPermissionNameNeed: 权限名称不能为空
      ErrorCode_AdminPermissionTypeInvalid: 权限类型无效
      ErrorCode_AdminRoleNotExist: 角色不存在
      ErrorCode_AuthTokenFailed: 鉴权失败
      ErrorCode_AuthTokenForbidden: 没有访问权限
      ErrorCode_AuthTokenInfoInvalid: Token信息无效
      ErrorCode_AuthTokenInspectInvalid: BearerToken无效
      ErrorCode_AuthTokenInvalid: BearerToken无效
      ErrorCode_Error: 失败
      ErrorCode_ReadContextRequestBodyFailed: 获取请求体内容失败
      ErrorCode_RecordNValidCanNotDeleted: 生效中的记录不能删除
      ErrorCode_RecordNotExist: 记录不存在
      ErrorCode_RequestBodyInvalid: 请求体参数无效
      ErrorCode_RequestParamsInvalid: 请求参数无效
      ErrorCode_RequestQueryInvalid: 请求查询参数无效
      ErrorCode_Success: 成功
    x-enum-varnames:
    - ErrorCode_Success
    - ErrorCode_Error
    - ErrorCode_ReadContextRequestBodyFailed
    - ErrorCode_RecordNotExist
    - ErrorCode_RecordNValidCanNotDeleted
    - ErrorCode_AuthTokenFailed
    - ErrorCode_AuthTokenInvalid
    - ErrorCode_AuthTokenInspectInvalid
    - ErrorCode_AuthTokenInfoInvalid
    - ErrorCode_AuthTokenForbidden
    - ErrorCode_RequestBodyInvalid
    - ErrorCode_RequestQueryInvalid
    - ErrorCode_RequestParamsInvalid
    - ErrorCode_AdminAccountPasswordInvalid
    - ErrorCode_AdminAccountNotExist
    - ErrorCode_AdminAccountInvalid
    - ErrorCode_AdminApiNameExist
    - ErrorCode_AdminApiPathExist
    - ErrorCode_AdminApiKeyExist
    - ErrorCode_AdminApiNotExist
    - ErrorCode_AdminPermissionKeyExist
    - ErrorCode_AdminPermissionExist
    - ErrorCode_AdminPermissionKeyNeed
    - ErrorCode_AdminPermissionNameNeed
    - ErrorCode_AdminPermissionTypeInvalid
    - ErrorCode_AdminMenuNotExist
    - ErrorCode_AdminRoleNotExist
info:
  contact: {}
paths:
...
  /admin/role/add:
    post:
      consumes:
      - application/json
      description: 添加角色
      parameters:
      - description: '@gotags: label:"角色描述" binding:"omitempty,trimBlank"'
        in: query
        name: describe
        type: string
      - description: '@gotags: label:"启用状态"'
        in: query
        name: enabled
        type: boolean
      - description: '@gotags: label:"角色名称" binding:"required,trimBlank"'
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/code.Message'
            - properties:
                data:
                  type: object
              type: object
      summary: 添加角色
      tags:
      - 角色接口相关
...
  /admin/role/bind_permissions:
    post:
      consumes:
      - application/json
      description: 角色绑定权限
      parameters:
      - description: '@gotags: label:"角色ID" binding:"required,min=1"'
        in: query
        minimum: 1
        name: id
        required: true
        type: integer
      - collectionFormat: csv
        description: '@gotags: label:"权限ID集合" binding:"required,min=1"'
        in: query
        items:
          type: integer
        minItems: 1
        name: permissionIds
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/code.Message'
            - properties:
                data:
                  type: object
              type: object
      summary: 角色绑定权限
      tags:
      - 角色接口相关
  /admin/role/delete:
    post:
      consumes:
      - application/json
      description: 删除角色
      parameters:
      - description: '@gotags: label:"角色ID" binding:"required,min=1"'
        in: query
        minimum: 1
        name: id
        required: true
        type: integer
      - collectionFormat: csv
        description: '@gotags: label:"权限ID集合" binding:"required,min=1"'
        in: query
        items:
          type: integer
        minItems: 1
        name: permissionIds
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/code.Message'
            - properties:
                data:
                  type: object
              type: object
      summary: 删除角色
      tags:
      - 角色接口相关
  /admin/role/edit:
    post:
      consumes:
      - application/json
      description: 编辑角色
      parameters:
      - description: '@gotags: label:"角色描述" binding:"omitempty,trimBlank"'
        in: query
        name: describe
        type: string
      - description: '@gotags: label:"角色ID" binding:"required,min=1"'
        in: query
        minimum: 1
        name: id
        required: true
        type: integer
      - description: '@gotags: label:"启用状态"'
        in: query
        name: isEnabled
        type: boolean
      - description: '@gotags: label:"角色名称" binding:"required,trimBlank"'
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/code.Message'
            - properties:
                data:
                  type: object
              type: object
      summary: 编辑角色
      tags:
      - 角色接口相关
  /admin/role/enable:
    post:
      consumes:
      - application/json
      description: 角色启用|禁用
      parameters:
      - description: '@gotags: label:"启用状态"'
        in: query
        name: enabled
        type: boolean
      - description: '@gotags: label:"角色ID" binding:"required,min=1"'
        in: query
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/code.Message'
            - properties:
                data:
                  type: object
              type: object
      summary: 角色启用|禁用
      tags:
      - 角色接口相关
  /admin/role/info:
    post:
      consumes:
      - application/json
      description: 角色详情
      parameters:
      - description: '@gotags: label:"角色ID" binding:"required,min=1"'
        in: query
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/code.Message'
            - properties:
                data:
                  $ref: '#/definitions/admin_proto.RespAdminRoleInfoData'
              type: object
      summary: 角色详情
      tags:
      - 角色接口相关
  /admin/role/list:
    post:
      consumes:
      - application/json
      description: 角色列表
      parameters:
      - description: 启用状态，0：全部，1：启用，2：禁用
        in: query
        name: enabled
        type: integer
      - description: 角色ID
        in: query
        name: id
        type: integer
      - description: 角色名称
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/code.Message'
            - properties:
                data:
                  $ref: '#/definitions/admin_proto.RespAdminRoleListData'
              type: object
      summary: 角色列表
      tags:
      - 角色接口相关
  /admin/role/permissions:
    post:
      consumes:
      - application/json
      description: 角色权限列表
      parameters:
      - description: '@gotags: label:"角色ID" binding:"required,min=1"'
        in: query
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/code.Message'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/admin_proto.RolePermissionItem'
                  type: array
              type: object
      summary: 角色权限列表
      tags:
      - 角色接口相关
...
swagger: "2.0"